Adds context to pipelines, changes handlers to async
This design decision was made as having 2-way data flow pipes breaks the isolation. Pipes have to know what data they have to return. This is impractical when you want to build reusable pipe, they no longer have the plug and play effect. Removing the coupling between pipes and responses allows them to be pretty much plug and play with the aid of mutators